Update setMetadata--> set_metadata
PyMuPDF had some major updates. It seems from their source code that setMetadata should still work as an alias, but I have version 1.21.0 of fitz/pymupdf and I was getting an error at this line. Changing to set_metadata fixes it. 

This could be something to submit as an issue to PyMuPDF rather than fixing here for backwards compatibility with old versions of PyMuPDF, but I didn't dig too deep on the PyMuPDF side. Leaving my fix here so I have it and in case you want it / it's useful to incorporate.